Update task handling of hypervisor cluster create
Update the polling of a hypervisor cluster create task.  In the case of
hypervisor create, the `sourceResourceUri` field in the final task response
points to a `system` and not the hypervisor cluster that has just been created.
(This differs from the task behaviour for datastores.)

Switch to instead extract the hypervisor cluster uuid from the returned task
body's `associatedResources` list. (Note that the `associatedResources` list
for datastores is empty, so we cannot consistently use this approach.)

```
  "associatedResources": [
    {
      "name": "cluster1",
      "resourceUri": "/virtualization/v1beta1/hypervisor-clusters/298a299e-78f5-5acb-86ce-4e9fdc290ab7",
      "type": "hypervisor-cluster"
    }
  ]

```

This change avoids terraform erroring after a successful hypervisor create.