Temporarily hack hypervisor delete SDK code
There is currently a bug where the service needs the remove
hypervisor cluster POST request to use snake case for the
id field, eg:

```
{"hypervisor_cluster_ids":["50af582d-3b98-570f-b113-650da9d0d633"]}
```

This should not be the case since the OpenAPI spec uses camelCase.
For now switch to use snake_case. Once the service is updated to conform
to the OpenAPI specification we will revert this change.