Add esxIpAddress to server
We add a new field `esxIpAddress` to the server model.
This field is an outlier - it does not fit cleanly into a declarative
model.

The field is used as part of an `add-hypervisor-server` RPC-style call.
It is not part of (and never will be part of) the newly created
server's attributes.

My understanding is that `esxIpAddress` will typically be set to something like
an ephemeral DHCP address that exists only when a server is first racked.  Once
add-hypervisor-server has completed, the newly created server's IP address will
have been changed to the `hypervisor_host.hypervisor_host_ip` value. This value
will be returned as `hypervisorHost.hypervisorHostIp` in subsequent http GET
requests to the relevant `/servers/<id>` endpoint.

The `esxIpAddress` value will be retained as part of the terraform
configuration for the server, however once a server has been created
this value is no longer of interest.