Allow supplying client id/secret for auth
Allow a user to specify a PCBe Client ID and PCBe Client
Secret as an alternative to a token.

Either (1) a token or (2) credentials (id/secret) must be provided.
The code will use the creds to generate a token when a token is
not provided directly.

Currently the environment variables being used are
`PCBE_CLIENT_ID` and `PCBE_CLIENT_SECRET`. These are defined
as constants and may be changed later (eg when integrating with
the "umbrella" hpegl provider) if desired.