Consume updated datastore model
The datastore terraform must now specify cluster_info id:

```
resource "hpegl_pc_datastore" "my_datastore" {
        name = "mclaren-ds19"
        datastore_type = "VMFS"
        capacity_in_bytes = 17179869184
        hci_cluster_uuid = data.hpegl_pc_system.my_system.id
        cluster_info = {
                "id": hpegl_pc_hypervisor_cluster.my_hypervisor_cluster.id
        }
}
```

This is to allow future changes where the existence of a datastore
will be inferred by filtering on datastore name, hci_cluster_uuid,
and cluster_info.id. The reasons for this are subtle, it relates
to `terraform apply` relying on PCBE operations; during an operation
the ID of the datastore is unknown. It is therefore important
for the triplet of name/hci_cluster_uuid/cluster_info.id to be
unique.